refactor(assets): name the mascot a mascot, not a logo
PEANUTMAN_LOGO was the peanut-character illustration mislabeled as a
"logo" — distinct from the app icon (peanut-icon.svg) and the PEANUT
wordmark (PEANUT_LOGO). The name described where it was first used (a
header), not what it depicts.

- Rename export PEANUTMAN_LOGO -> PEANUTMAN and file
  mascot/peanut-logo.svg -> mascot/peanutman.svg; update all call sites,
  the local ASSET_ var, and the mascot alt text ("logo" -> "Peanut mascot").
- Move PEANUT_LOGO_BLACK out of illustrations/ into logos/ (it's the brand
  logo); fix the one subpath import.

No behavior change — same pixels, accurate names. public/peanutman-logo.svg
is retained unchanged as the external email/embed alias.
