+++ src/app/(mobile-ui)/qr-pay/__tests__/qr-pay-states.test.tsx
@@ -81,7 +81,7 @@ jest.mock('@/assets/payment-apps', () => ({
 
 // The page imports PeanutThinking from @/assets/mascot and STAR_STRAIGHT_ICON
 // from @/assets/icons directly — mock those paths, not the @/assets barrel, and
-// keep sibling exports (e.g. PEANUTMAN_LOGO, ETHEREUM_ICON used by QRScanner) intact.
+// keep sibling exports (e.g. PEANUTMAN, ETHEREUM_ICON used by QRScanner) intact.
 jest.mock('@/assets/mascot', () => ({
     ...jest.requireActual('@/assets/mascot'),
     PeanutThinking: '/peanut-guy.gif',
