+++ src/assets/illustrations/index.ts
@@ -8,7 +8,6 @@ export { default as HandThumbsUpV2 } from './hand-thumbs-up-v2.svg'
 export { default as HandToken } from './new-hand-token.svg'
 export { default as AboutPeanut } from './new-hero-description.svg'
 //export { default as PeanutArmHoldingBeer } from './peanut-arm-holding-beer.svg'
-export { default as PEANUT_LOGO_BLACK } from './peanut-logo-dark.svg'
 export { default as PeanutsBG } from './peanuts-bg.svg'
 export { default as Sparkle } from './sparkle.svg'
 export { default as Star } from './star.svg'
