+++ src/assets/logos/index.ts
@@ -1,4 +1,5 @@
 export { default as PEANUT_LOGO } from './peanut-logo.svg';
+export { default as PEANUT_LOGO_BLACK } from './peanut-logo-dark.svg';
 export { default as MEPA_ARGENTINA_LOGO } from './mepa-arg.svg';
 export { default as PIX_BRZ_LOGO } from './pix-brz.svg';
 export { default as DEVCONNECT_LOGO } from './devconnect.svg';
