+++ src/assets/mascot/index.ts
@@ -15,7 +15,7 @@ export { default as PeanutWalking } from './peanut-walking.webp' // mid-stride, 
 export { default as PeanutPointingDown } from './peanut-pointing-down.webp' // both hands pointing down — marketing CTA
 
 // Stills
-export { default as PEANUTMAN_LOGO } from './peanut-logo.svg'
+export { default as PEANUTMAN } from './peanutman.svg'
 export { default as PEANUTMAN_PFP } from './peanut-pfp.svg'
 export { default as PEANUTMAN_HOLDING_BEER } from './peanut-beer.svg'
 export { default as PEANUTMAN_MOBILE } from './peanut-club.webp'
