+++ src/components/Global/ConfirmInviteModal/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { type FC } from 'react'
 import Image from 'next/image'
-import { PEANUT_LOGO_BLACK, PEANUTMAN_LOGO } from '@/assets'
+import { PEANUT_LOGO_BLACK, PEANUTMAN } from '@/assets'
 import Modal from '../Modal'
 import { Button } from '@/components/0_Bruddle/Button'
 import { PeanutWavingHello } from '@/assets/mascot'
@@ -52,7 +52,7 @@ const ConfirmInviteModal: FC<ConfirmInviteModalProps> = ({
                         <Button className="w-full" shadowSize="4" variant="purple" onClick={handleContinueWithPeanut}>
                             <div>Join</div>
                             <div className="flex items-center gap-1">
-                                <Image src={PEANUTMAN_LOGO} alt="Peanut Logo" className="size-5" />
+                                <Image src={PEANUTMAN} alt="Peanut Logo" className="size-5" />
                                 <Image src={PEANUT_LOGO_BLACK} alt="Peanut Logo" />
                             </div>
                         </Button>
